requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
langgraph>=0.2.46
langchain>=0.1.0
langchain-openai>=0.0.3
sentence-transformers>=2.2.2
faiss-cpu==1.12.0
redis==5.0.1
sqlite-utils==3.35.2
python-multipart==0.0.6
aiofiles==23.2.0
httpx==0.25.2
python-dotenv==1.0.0
loguru==0.7.2